RESKC-1953: Award Budget Personnel Screen Locks Up when You add a Summary Line Item (#2156)
Steps to Recreate: On award budget document, navigate to the Personnel page and attempt to add a line item with the Person "Summary". I confirmed this with Object Codes Grad Assistants and Dual Non Teaching, but when I chose a person the line added properly.
Result: The header for the line that you just tried to add will load, but nothing on the screen after that will load.