RESKC-2238: Fixing cost share issue. (#2341)
When a user has a budget with cost share and the cost share distribution has multiple rows with SAME source in SAME FY; then the following error results: Row duplication!! Please enter a different Fiscal Year or Source Account for 1 or 1234567
If the user was somehow able to save this information and then tries to access the budget back again, this error comes up when the budget is opened, however, it prevents the user from navigating to the right section to fix the error.
The expected behavior is that the error is present, BUT the user should be able to navigate between the section and to the right section to fix it. The error should not let user set the budget as Complete.
New Message:
Multiple rows with the same Fiscal Year and Source Account combination are not allowed. Please fix.