RESKC-2228: ignoring schema failures for the NIH service.
There isn't a good way to know if opportunities are supported by the NIH validation service.  We are already doing schema validation internally so any schema validation failure likely indicates that the opportunity contains forms that NIH doesn't understand.  We detect this condition and downgrade the message to info level and display that NIH validation is not active.