RESKC-2021: Fixing rules so that argValue lookup values are checked. (#2383)
In the IP, Award and Subaward modules, on the Custom Data tab you can either lookup and return custom data arg values or manually type values directly into each custom data field. Manually entering data into these fields allows ANY arg lookup value to be used regardless if it's a returnable value for that specific field. This does not throw an error and allows submitting to final status which could lead to major data integrity issues.
For example - in res-demo1:
In the Subaward Module:
I have a My Custom Data field that uses yes_no_flag Argument Name HOWEVER, I can enter "Kat Szulc" for My Custom Data even-though that is NOT a returnable value
It appears that any arg value option on ANY of the fields is allowed regardless if it's an option on the returnable list of values.