RESKC-2322: Fix subaward Estimated Project Period End Date. (#2401)
SUMMARY: In the FDP Cost Reimbursement Research Subaward Agreement printout the Estimated Project Period section should be mapping the Start Date of the last Financial entry and the End Date of the over Project End Date from the Subaward tab. Currently the End Date is also mapping from the Financial entry. The Estimated End Date needs to be updated to map from the Subaward End Date.
STEPS TO REPRODUCE:
create Subaward
fill out required fields for saving - input Start/End Date on the Subaward tab (i.e. 6/1/17 - 6/30/17)
go to the Financial tab and add transaction entry with Start/End Date of 6/2/17-6/29/17
fill out required info on the Templates tab
go to Subaward Actions and print the FDP Subaward Agreement
notice the Estimated Project Period End Date is 6/29/18 when it should be 6/30/18 from the Project End Date on the Subaward tab.