RESKC-2232: Enabling notifications (#2692)
Steps to Reproduce
Create an award with basic information to submit/finalize
Add a property report with a due date in the past
submit award
Actual Result
No notification is sent out to the persons listed on the notification document.
Expected Result
The system should generate emails/action list items triggered off of the events specified in the notifications table. (Ex: Property report Overdue) It looks like the triggers are not in place for these to be sending.