RESKC-2729: Clean up report tracking data that violates referential integrity
These scripts clean up report tracking data that was unable to to be re-mapped to the report terms associated with the latest award versions as part of commit 37e771ee9b066fc541a2bb8c39952adbf05feecd. A lot of this data was likely duplicated due to consistency issues in legacy report tracking code, such as tracking data being incorrectly copied across versions and failing to properly update report tracking properties when the associated award and award term properties were changed. Report tracking data that was unable to be adapted to the new consistency constraints will be moved to a new `award_report_tracking_backup` table. That way these entries can be examined to see if they need to be manually re-associated with award report terms or if they can be safely ignored/deleted.